Refactor QR code handling: Move ScanQR and LoginByQR components to features/qrScan, update imports, and enhance socket connection management.
